Develop/v3

+++ controllers/auth/login.auth.controller.js
@@ -5,6 +5,7 @@ const tokenService = require("../../services/auth/auth.token.service");
 const { encrypt62 } = require("../../utils/encrypt.util");
 
 const { logError } = require("../../utils/handlers/error.logger");
+const registerService = require("../../services/auth/register.auth.service");
 
 const localLogin = async (req, res, next) => {
   try {
@@ -35,16 +36,20 @@ const localLogin = async (req, res, next) => {
       \nRT : ${refreshToken}`
     );
 
+    const ret = {
+      user_id: encrypt62(data.user_id),
+      name: data.name,
+      nickname: data.nickname,
+    };
+
+    logger.debug(`[localLogin] 응답 데이터: ${JSON.stringify(ret, null, 2)}`);
+
     // 5. 응답
     return res
       .status(200)
       .cookie("SPECTOGETHER_RT", refreshToken, refreshTokenCookieOptions)
       .success({
-        user: {
-          user_id: encrypt62(data.user_id),
-          name: data.name,
-          nickname: data.nickname,
-        },
+        user: ret,
         access_token: accessToken,
       });
   } catch (err) {
@@ -58,17 +63,14 @@ const logout = async (req, res, next) => {
     // 1. RT가 DB에 존재하는지 확인하고, 존재한다면 DB에서 제거
     // 1-1. DB에 존재하지 않으면 에러 로깅
     const refreshToken = req.cookies.SPECTOGETHER_RT;
-    tokenService.checkIfRefreshTokenExists(refreshToken);
+    const isTokenExist = tokenService.checkIfRefreshTokenExists(refreshToken);
     // 1-2. DB에서 제거
-    await loginService.deleteRefreshToken(refreshToken);
+    if (isTokenExist) await loginService.deleteRefreshToken(refreshToken);
 
     // 2. 응답
-    return res
-      .status(200)
-      .clearCookie("SPECTOGETHER_RT", refreshTokenCookieOptions)
-      .success({
-        message: "로그아웃 되었습니다.",
-      });
+    return res.status(200).clearCookie("SPECTOGETHER_RT").success({
+      message: "로그아웃 되었습니다.",
+    });
   } catch (err) {
     logError(err);
     next(err);
@@ -101,8 +103,62 @@ const reissueAccessToken = async (req, res, next) => {
   }
 };
 
+const resetPassword = async (req, res, next) => {
+  try {
+    // 1. 들어온 입력 검증
+    // 2. 해당 세션의 전화번호 찾기
+    const phone = await registerService.getPhoneNumberByVerificationId(
+      req.body.phone_verification_session_id
+    );
+    // 3. 해당 전화번호를 가진 사용자의 비밀번호 변경
+    const data = await loginService.setPassword({
+      phone: phone,
+      newPassword: req.body.new_password,
+    });
+
+    // 4. 응답
+    return res.status(200).success({
+      message: "비밀번호가 초기화되었습니다.",
+    });
+  } catch (err) {
+    logError(err);
+    next(err);
+  }
+};
+
+const changePassword = async (req, res, next) => {
+  try {
+    // 1. 들어온 입력 검증
+    // 2. 해당 세션의 전화번호 찾기
+    const phone = await registerService.getPhoneNumberByVerificationId(
+      req.body.phone_verification_session_id
+    );
+    // 3. 기존 비밀번호가 일치하는지 확인
+    const oldData = await loginService.getUserPasswordByPhoneNumber(phone);
+    await loginService.comparePassword({
+      password: req.body.old_password,
+      hashed_password: oldData.password,
+    });
+    // 4. 일치하면 새로운 비밀번호로 변경
+    await loginService.setPassword({
+      phone: phone,
+      newPassword: req.body.new_password,
+    });
+
+    // 4. 응답
+    return res.status(200).success({
+      message: "비밀번호가 변경되었습니다.",
+    });
+  } catch (err) {
+    logError(err);
+    next(err);
+  }
+};
+
 module.exports = {
   localLogin,
   logout,
   reissueAccessToken,
+  resetPassword,
+  changePassword,
 };
